requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.1.2
django-crispy-forms==1.14.0
django-resized==1.0.2
django-tinymce==3.5.0
django-user-agents==0.4.0
gunicorn==20.1.0
kombu==5.2.4
mypy-extensions==0.4.3
pathspec==0.10.2
Pillow==9.2.0
platformdirs==2.5.4
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.7
PyYAML==6.0
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.4.0
ua-parser==0.16.1
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
